+++ src/main/java/im/aop/loggers/messageinterpolation/IterableToStringStrategy.java
@@ -19,7 +19,7 @@ public class IterableToStringStrategy implements ToStringStrategy {
   }
 
   @Override
-  public boolean canSupport(Class<?> type) {
+  public boolean supports(Class<?> type) {
     return Iterable.class.isAssignableFrom(type);
   }