+++ src/main/java/im/aop/loggers/messageinterpolation/OptionalToStringStrategy.java
@@ -18,7 +18,7 @@ public class OptionalToStringStrategy implements ToStringStrategy {
   }
 
   @Override
-  public boolean canSupport(Class<?> type) {
+  public boolean supports(Class<?> type) {
     return Optional.class.isAssignableFrom(type);
   }