+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- Rename canSupport() method to supports() in ToStringStrategy.
 
 ## [1.0.5] - 2022-01-16